feat(*): flexible duration of election phases
Add option to change duration of the phases inside election.
Restriction is, that total election duration must stay the same.

Propose phase is more time consuming for validators, because data needs to be prepared.
When we will be able to adjust propose/reveal ratio we can make more time for send a proposal.